fix: working appimage build and run.sh

+++ appimage-build.sh
@@ -5,8 +5,17 @@ BUILDCACHE="/tmp"
 EXEC="scc"
 LIB="lib"
 
-EVDEV_VERSION=0.7.0
+PYTHON=${PYTHON:-python3}
+PYVERSION=$($PYTHON -c 'import sys; print("%d.%d" % sys.version_info[:2])')
+
+EVDEV_VERSION=1.9.1
+PYLIBACL_VERSION=0.7.3
+
+ARCH_PYVERSION=3.14
+BUNDLED_PY=3.14.7
+
 [ x"$BUILD_APPDIR" == "x" ] && BUILD_APPDIR=$(pwd)/appimage
+SITE=usr/lib/python${PYVERSION}/site-packages
 
 
 function download_dep() {
@@ -29,11 +38,9 @@ function build_dep() {
 	mkdir -p ${BUILDCACHE}/${NAME}
 	pushd ${BUILDCACHE}/${NAME}
 	tar --extract --strip-components=1 -f ${DEPCACHE}/${NAME}.tar.gz
-	PYTHONPATH=${BUILD_APPDIR}/usr/lib/python2.7/site-packages python2 \
+	PYTHONPATH=${BUILD_APPDIR}/${SITE} ${PYTHON} \
 		setup.py install --optimize=1 \
 		--prefix="/usr/" --root="${BUILD_APPDIR}"
-	mkdir -p "${BUILD_APPDIR}/usr/lib/python2.7/site-packages/"
-	python2 setup.py install --prefix="/usr/" --root="${BUILD_APPDIR}"
 	popd
 }
 
@@ -41,56 +48,70 @@ function unpack_dep() {
 	NAME="$1"
 	pushd ${BUILD_APPDIR}
 	tar --extract --exclude="usr/include**" --exclude="usr/lib/pkgconfig**" \
-			--exclude="usr/lib/python3.6**" -f ${DEPCACHE}/${NAME}.tar.gz
+			--exclude="usr/lib/python2.7**" -f ${DEPCACHE}/${NAME}.tar.gz
 	popd
 }
 
 set -ex		# display commands, terminate after 1st failure
 
+# Verify host Python matches the bundled Arch Python
+if [ "$PYVERSION" != "$ARCH_PYVERSION" ] ; then
+	echo "ERROR: appimage-build.sh requires Python $ARCH_PYVERSION to build the AppImage,"
+	echo "but system Python reports $PYVERSION ($(which ${PYTHON}))."
+	echo "Install Python $ARCH_PYVERSION and set PYTHON to its interpreter."
+	exit 1
+fi
+
 # Download deps
-download_dep "python-2.7.18" "https://archive.archlinux.org/packages/p/python2/python2-2.7.18-1-x86_64.pkg.tar.zst"
-download_dep "python-evdev-0.7.0" "https://github.com/gvalkov/python-evdev/archive/v0.7.0.tar.gz"
-download_dep "pylibacl-0.5.3" "https://github.com/iustin/pylibacl/releases/download/pylibacl-v0.5.3/pylibacl-0.5.3.tar.gz"
-download_dep "python-gobject-3.26.1" "https://archive.archlinux.org/packages/p/python2-gobject/python2-gobject-3.26.1-1-x86_64.pkg.tar.xz"
-download_dep "python-cairo-1.18.2" "https://archive.archlinux.org/packages/p/python2-cairo/python2-cairo-1.18.2-3-x86_64.pkg.tar.xz"
-download_dep "gobject-introspection-runtime-1.60" "https://archive.archlinux.org/packages/g/gobject-introspection-runtime/gobject-introspection-runtime-1.60.0-1-x86_64.pkg.tar.xz"
-download_dep "libpng-1.6.34" "https://archive.archlinux.org/packages/l/libpng/libpng-1.6.34-2-x86_64.pkg.tar.xz"
-download_dep "gdk-pixbuf-2.36.9" "https://archive.archlinux.org/packages/g/gdk-pixbuf2/gdk-pixbuf2-2.36.9-1-x86_64.pkg.tar.xz"
-download_dep "libffi-3.3" "https://archive.archlinux.org/packages/l/libffi/libffi-3.3-4-x86_64.pkg.tar.zst"
-download_dep "libcroco-0.6.13" "https://archive.archlinux.org/packages/l/libcroco/libcroco-0.6.13-1-x86_64.pkg.tar.xz"
-download_dep "libxml2-2.9.7" "https://archive.archlinux.org/packages/l/libxml2/libxml2-2.9.7%2B4%2Bg72182550-2-x86_64.pkg.tar.xz"
-download_dep "librsvg-2.44.12" "https://archive.archlinux.org/packages/l/librsvg/librsvg-2%3A2.44.12-1-x86_64.pkg.tar.xz"
-download_dep "libffi-3.2.1" "https://archive.archlinux.org/packages/l/libffi/libffi-3.2.1-4-x86_64.pkg.tar.xz"
-download_dep "icu-60.2" "https://archive.archlinux.org/packages/i/icu/icu-60.2-1-x86_64.pkg.tar.xz"
-download_dep "zlib-1:1.2.12" "https://archive.archlinux.org/packages/z/zlib/zlib-1%3A1.2.12-2-x86_64.pkg.tar.zst"
+download_dep "python-${BUNDLED_PY}" "https://archive.archlinux.org/packages/p/python/python-${BUNDLED_PY}-1-x86_64.pkg.tar.zst"
+download_dep "python-evdev-${EVDEV_VERSION}" "https://github.com/gvalkov/python-evdev/archive/refs/tags/v${EVDEV_VERSION}.tar.gz"
+download_dep "pylibacl-${PYLIBACL_VERSION}" "https://files.pythonhosted.org/packages/cd/9e/e23f907c8e2cdc721c3d87eddda0cedee2f7cd7edf22f8439cee67f48a03/pylibacl-${PYLIBACL_VERSION}.tar.gz"
+download_dep "python-gobject-3.58.0" "https://archive.archlinux.org/packages/p/python-gobject/python-gobject-3.58.0-1-x86_64.pkg.tar.zst"
+download_dep "python-cairo-1.29.1" "https://archive.archlinux.org/packages/p/python-cairo/python-cairo-1.29.1-1-x86_64.pkg.tar.zst"
+download_dep "gobject-introspection-runtime-1.86.0" "https://archive.archlinux.org/packages/g/gobject-introspection-runtime/gobject-introspection-runtime-1.86.0-2-x86_64.pkg.tar.zst"
+download_dep "gdk-pixbuf2-2.44.7" "https://archive.archlinux.org/packages/g/gdk-pixbuf2/gdk-pixbuf2-2.44.7-1-x86_64.pkg.tar.zst"
+download_dep "librsvg-2.62.91" "https://archive.archlinux.org/packages/l/librsvg/librsvg-2%3A2.62.91-1-x86_64.pkg.tar.zst"
+download_dep "cairo-1.18.4" "https://archive.archlinux.org/packages/c/cairo/cairo-1.18.4-1-x86_64.pkg.tar.zst"
+download_dep "libpng-1.6.58" "https://archive.archlinux.org/packages/l/libpng/libpng-1.6.58-2-x86_64.pkg.tar.zst"
+download_dep "icu-78.3" "https://archive.archlinux.org/packages/i/icu/icu-78.3-1-x86_64.pkg.tar.zst"
+download_dep "zlib-1.3.2" "https://archive.archlinux.org/packages/z/zlib/zlib-1%3A1.3.2-3-x86_64.pkg.tar.zst"
 
 # Prepare & build deps
-export PYTHONPATH=${BUILD_APPDIR}/usr/lib/python2.7/site-packages/
+export PYTHONPATH=${BUILD_APPDIR}/${SITE}
 mkdir -p "$PYTHONPATH"
 if [[ $(grep ID_LIKE /etc/os-release) == *"suse"* ]] ; then
 	# Special handling for OBS
 	ln -s lib64 ${BUILD_APPDIR}/usr/lib
-	export PYTHONPATH="$PYTHONPATH":${BUILD_APPDIR}/usr/lib64/python2.7/site-packages/
+	export PYTHONPATH="$PYTHONPATH":${BUILD_APPDIR}/usr/lib64/python${PYVERSION}/site-packages/
 	LIB=lib64
 fi
 
-build_dep "python-evdev-0.7.0"
-build_dep "pylibacl-0.5.3"
-unpack_dep "python-2.7.18"
-unpack_dep "libpng-1.6.34"
-unpack_dep "python-cairo-1.18.2"
-unpack_dep "libffi-3.2.1"
-unpack_dep "python-gobject-3.26.1"
-unpack_dep "gobject-introspection-runtime-1.60"
-unpack_dep "gdk-pixbuf-2.36.9"
-unpack_dep "libffi-3.3"
-unpack_dep "libcroco-0.6.13"
-unpack_dep "libxml2-2.9.7"
-unpack_dep "librsvg-2.44.12"
-unpack_dep "icu-60.2"
-unpack_dep "zlib-1:1.2.12"
+build_dep "python-evdev-${EVDEV_VERSION}"
+build_dep "pylibacl-${PYLIBACL_VERSION}"
+unpack_dep "python-${BUNDLED_PY}"
+unpack_dep "libpng-1.6.58"
+unpack_dep "python-cairo-1.29.1"
+unpack_dep "python-gobject-3.58.0"
+unpack_dep "gobject-introspection-runtime-1.86.0"
+unpack_dep "gdk-pixbuf2-2.44.7"
+unpack_dep "cairo-1.18.4"
+unpack_dep "librsvg-2.62.91"
+unpack_dep "icu-78.3"
+unpack_dep "zlib-1.3.2"
 
-# Remove uneeded files
+# Verify bundled Python runs and matches the host ABI
+PYTHONHOME=${BUILD_APPDIR}/usr ${BUILD_APPDIR}/usr/bin/python${PYVERSION} \
+	-c 'import sys; print("Bundled Python OK:", sys.version.split()[0])'
+HOST_SUFFIX=$(${PYTHON} -c 'import importlib.machinery; print(importlib.machinery.EXTENSION_SUFFIXES[0])')
+BUNDLED_SUFFIX=$(PYTHONHOME=${BUILD_APPDIR}/usr ${BUILD_APPDIR}/usr/bin/python${PYVERSION} \
+	-c 'import importlib.machinery; print(importlib.machinery.EXTENSION_SUFFIXES[0])')
+if [ "$HOST_SUFFIX" != "$BUNDLED_SUFFIX" ] ; then
+	echo "ERROR: host Python extension suffix '$HOST_SUFFIX' does not match bundled"
+	echo "Python extension suffix '$BUNDLED_SUFFIX'. C modules would not load."
+	exit 1
+fi
+
+# Remove unneeded files
 rm -f "${BUILD_APPDIR}/usr/${LIB}/gdk-pixbuf-2.0/2.10.0/loaders/libpixbufloader-ani.so"
 rm -f "${BUILD_APPDIR}/usr/${LIB}/gdk-pixbuf-2.0/2.10.0/loaders/libpixbufloader-bmp.so"
 rm -f "${BUILD_APPDIR}/usr/${LIB}/gdk-pixbuf-2.0/2.10.0/loaders/libpixbufloader-gif.so"
@@ -101,24 +122,24 @@ rm -f "${BUILD_APPDIR}/usr/${LIB}/gdk-pixbuf-2.0/2.10.0/loaders/libpixbufloader-
 rm -f "${BUILD_APPDIR}/usr/${LIB}/gdk-pixbuf-2.0/2.10.0/loaders/libpixbufloader-qtif.so"
 rm -f "${BUILD_APPDIR}/usr/${LIB}/gdk-pixbuf-2.0/2.10.0/loaders/libpixbufloader-tga.so"
 rm -f "${BUILD_APPDIR}/usr/${LIB}/gdk-pixbuf-2.0/2.10.0/loaders/libpixbufloader-tiff.so"
-rm -R "${BUILD_APPDIR}/usr/lib/cmake"
-rm -R "${BUILD_APPDIR}/usr/share/doc"
-rm -R "${BUILD_APPDIR}/usr/share/gtk-doc"
-rm -R "${BUILD_APPDIR}/usr/share/locale"
-rm -R "${BUILD_APPDIR}/usr/share/man"
-rm -R "${BUILD_APPDIR}/usr/share/thumbnailers"
-rm -R "${BUILD_APPDIR}/usr/share/vala"
-rm -R "${BUILD_APPDIR}/usr/share/icu"
+rm -Rf "${BUILD_APPDIR}/usr/lib/cmake"
+rm -Rf "${BUILD_APPDIR}/usr/share/doc"
+rm -Rf "${BUILD_APPDIR}/usr/share/gtk-doc"
+rm -Rf "${BUILD_APPDIR}/usr/share/locale"
+rm -Rf "${BUILD_APPDIR}/usr/share/man"
+rm -Rf "${BUILD_APPDIR}/usr/share/thumbnailers"
+rm -Rf "${BUILD_APPDIR}/usr/share/vala"
+rm -Rf "${BUILD_APPDIR}/usr/share/icu"
 
 # Build important part
-python2 setup.py build
-python2 setup.py install --prefix ${BUILD_APPDIR}/usr
+${PYTHON} setup.py build
+PYTHONPATH=${BUILD_APPDIR}/${SITE} ${PYTHON} setup.py install --prefix ${BUILD_APPDIR}/usr
 
 # Move udev stuff
 mv ${BUILD_APPDIR}/usr/lib/udev/rules.d/69-${APP}.rules ${BUILD_APPDIR}/
 rmdir ${BUILD_APPDIR}/usr/lib/udev/rules.d/
 rmdir ${BUILD_APPDIR}/usr/lib/udev/
-cp "/usr/include/linux/input-event-codes.h" ${BUILD_APPDIR}/usr/${LIB}/python2.7/site-packages/scc/
+cp "/usr/include/linux/input-event-codes.h" ${BUILD_APPDIR}/usr/${LIB}/python${PYVERSION}/site-packages/scc/
 
 # Move & patch desktop file
 mv ${BUILD_APPDIR}/usr/share/applications/${APP}.desktop ${BUILD_APPDIR}/
@@ -134,7 +155,7 @@ cp scripts/${APP}.appdata.xml ${BUILD_APPDIR}/usr/share/metainfo/${APP}.appdata.
 
 # Fix shebangs
 for x in "${BUILD_APPDIR}/usr/bin"/sc-controller "${BUILD_APPDIR}/usr/bin"/scc* ; do
-	sed -i 's|#!/usr/bin/python2|#!/usr/bin/env python2|' "$x"
+	sed -i 's|^#!.*python2.*|#!/usr/bin/env python3|' "$x"
 done
 
 # Copy AppRun script
